Converted ChatApiContext to TS and removed legacy code
